Allow sample data generation on staging website
Checks are used in scripts to prevent sample data being added
on production website.

An alternative system was investigated that uses different
Dockerfiles for updating data, however this system was very
complicated for preventing dependencies for creating
sample data from existing in the production environment.

Including the dependencies in all versions allows for a much
simplier system for contributors to understand.